+++ packages/sdk/src/encryption/PublisherKeyExchange.ts
@@ -1,4 +1,12 @@
-import { EthereumAddress, Logger, StreamPartID, StreamPartIDUtils, UserID } from '@streamr/utils'
+import {
+    EthereumAddress,
+    Logger,
+    StreamPartID,
+    StreamPartIDUtils,
+    UserID,
+    isEthereumAddressUserId,
+    toEthereumAddress
+} from '@streamr/utils'
 import without from 'lodash/without'
 import { Lifecycle, inject, scoped } from 'tsyringe'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
@@ -67,10 +75,13 @@ export class PublisherKeyExchange {
         })
         eventEmitter.on('messagePublished', (msg) => {
             if (msg.signatureType === SignatureType.ERC_1271) {
-                const address = msg.getPublisherId()
-                if (!this.erc1271ContractAddresses.has(address)) {
-                    logger.debug('Add ERC-1271 publisher', { address })
-                    this.erc1271ContractAddresses.add(address)
+                const publisherId = msg.getPublisherId()
+                if (isEthereumAddressUserId(publisherId)) {
+                    const address = toEthereumAddress(publisherId)
+                    if (!this.erc1271ContractAddresses.has(address)) {
+                        logger.debug('Add ERC-1271 publisher', { address })
+                        this.erc1271ContractAddresses.add(address)
+                    }
                 }
             }
         })
@@ -84,7 +95,7 @@ export class PublisherKeyExchange {
                 if (responseType !== ResponseType.NONE) {
                     this.logger.debug('Handling group key request', { requestId, responseType })
                     await validateStreamMessage(request, this.streamRegistry, this.signatureValidator)
-                    const authenticatedUser = await this.authentication.getAddress()
+                    const authenticatedUser = await this.authentication.getUserId()
                     const keys = without(
                         await Promise.all(groupKeyIds.map((id: string) => this.store.get(id, authenticatedUser))),
                         undefined) as GroupKey[]
@@ -116,11 +127,11 @@ export class PublisherKeyExchange {
         }
     }
 
-    private async getResponseType(publisher: UserID): Promise<ResponseType> {
-        const authenticatedUser = await this.authentication.getAddress()
-        if (publisher === authenticatedUser) {
+    private async getResponseType(publisherId: UserID): Promise<ResponseType> {
+        const authenticatedUser = await this.authentication.getUserId()
+        if (publisherId === authenticatedUser) {
             return ResponseType.NORMAL
-        } else if (this.erc1271ContractAddresses.has(publisher)) {
+        } else if (isEthereumAddressUserId(publisherId) && this.erc1271ContractAddresses.has(toEthereumAddress(publisherId))) {
             return ResponseType.ERC_1271
         } else {
             return ResponseType.NONE
@@ -130,10 +141,10 @@ export class PublisherKeyExchange {
     private async createResponse(
         keys: GroupKey[],
         responseType: ResponseType,
-        publisher: UserID,
+        publisherId: UserID,
         streamPartId: StreamPartID,
         rsaPublicKey: string,
-        recipient: UserID,
+        recipientId: UserID,
         requestId: string
     ): Promise<StreamMessage> {
         const encryptedGroupKeys = await Promise.all(keys.map((key) => {
@@ -141,7 +152,7 @@ export class PublisherKeyExchange {
             return new EncryptedGroupKey(key.id, encryptedGroupKey)
         }))
         const responseContent = new OldGroupKeyResponse({
-            recipient,
+            recipient: recipientId,
             requestId,
             encryptedGroupKeys
         })
@@ -151,7 +162,7 @@ export class PublisherKeyExchange {
                 StreamPartIDUtils.getStreamPartition(streamPartId),
                 Date.now(),
                 0,
-                publisher,
+                publisherId,
                 createRandomMsgChainId()
             ),
             content: convertGroupKeyResponseToBytes(responseContent),