+++ packages/sdk/test/end-to-end/Permissions.test.ts
@@ -1,12 +1,12 @@
 import { Wallet } from 'ethers'
 
-import { createRelativeTestStreamId } from '../test-utils/utils'
-import { CONFIG_TEST } from '../../src/ConfigTest'
-import { StreamrClient } from '../../src/StreamrClient'
-import { Stream } from '../../src/Stream'
-import { StreamPermission } from '../../src/permission'
 import { fastWallet, fetchPrivateKeyWithGas, randomUserId } from '@streamr/test-utils'
-import { toEthereumAddress } from '@streamr/utils'
+import { toUserId } from '@streamr/utils'
+import { CONFIG_TEST } from '../../src/ConfigTest'
+import { Stream } from '../../src/Stream'
+import { StreamrClient } from '../../src/StreamrClient'
+import { StreamPermission } from '../../src/permission'
+import { createRelativeTestStreamId } from '../test-utils/utils'
 
 const TIMEOUT = 40000
 
@@ -114,7 +114,7 @@ describe('Stream permissions', () => {
     it('get permissions', async () => {
         await stream.grantPermissions({ public: true, permissions: [StreamPermission.PUBLISH] })
         const permissions = await stream.getPermissions()
-        const owner = await client.getAddress()
+        const owner = await client.getUserId()
         return expect(permissions).toIncludeSameMembers([{
             user: owner,
             permissions: [
@@ -220,7 +220,7 @@ describe('Stream permissions', () => {
         const message = {
             foo: Date.now()
         }
-        const errorSnippet = `You don't have permission to publish to this stream. Using address: ${toEthereumAddress(otherUser.address)}`
+        const errorSnippet = `You don't have permission to publish to this stream. Using address: ${toUserId(otherUser.address)}`
         await expect(() => otherUserClient.publish(stream.id, message)).rejects.toThrow(errorSnippet)
         await client.grantPermissions(stream.id, {
             user: otherUser.address,