+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Changes before Tatum release are not documented in this file.
 
 #### Changed
 
+- **BREAKING CHANGE:** Rename `getAddress()` to `getUserId()` (https://github.com/streamr-dev/network/pull/2774)
+
 #### Deprecated
 
 #### Removed